requirements.txt
python-telegram-bot[job-queue]==20.7
asyncpg==0.29.0
httpx~=0.25.2
python-dotenv==1.0.0
